Create Counter view
For now, it has the ability to set itself online or offline.
It can call new queues that have the 'Waiting' state.
It can set the called queue's state to either:
    - back to 'Waiting'
    - set to 'Completed'
    - set to 'Cancelled'
It now utilizes /api/queues/{queue} endpoints for updating queue status.
Removed setWaitingApi, setCompletedApi, and setCancelledApi from CounterController.php.

Later implementations should look to add an employee_id foreignKey to Counters. This being null would signify a counter being 'offline'.

'On Break' should probably also be a state. This should allow the cashier to stay logged in but allow the counter to pause from being able to take in customers.